DDA-86:  Initial .forEach loop for converting to relative paths + snapshot spec can be run
