+++ electron/channels/generate-diograph-channel.js
@@ -9,6 +9,19 @@ exports.generateDiographEventHandler = async function generateDiographEventHandl
   const diograph = existingDiograph
     ? compareAndMergeDiographs(existingDiograph, folderStructureDiograph)
     : folderStructureDiograph
+
+  // Relative paths for image & contentUrl
+  // TODO: Move to own function, how to make immutable?
+  Object.keys(diograph.diograph).forEach((dioryId) => {
+    const diory = diograph.diograph[dioryId]
+    if (diory.image) {
+      diory.image = diory.image.replace(`${path}/`, '')
+    }
+    if (diory.data && diory.data.contentUrl) {
+      diory.data = diory.data.contentUrl.replace(`${path}/`, '')
+    }
+  })
+
   await saveDiographJson(path, diograph.diograph, diograph.rootId)
   return { ...diograph, path }
 }
